(PUP-2794) Update Callable Type specification
PUP-2794 introduces the ability to specify a Callable with Union
type (ignored type). The specification only shows the result of this
internal use of Union, the Union type is not documented as it is an
implementation detail.

This updates the specification as the type algebra of Callable
has changed in accordance with PUP-2794.