(PUP-3001) Specify Resource Expression Operations
In PUP-3001, the change of not using dynamic scoping for
resource defaults was reversed, and this triggered the need to update
the specification (which had no text for this).

This commit also fixes unfinished parts of the specification in
modus-operandi, moves one section from catalog_expressions to 
modus_operandi (about auto loading).