(maint) Cleanup introduction to type system
The introduction to the type system repeated the same information twice
with a small variation. This confused readers.

It is also clarified that the names "Data Types" etc. are just
conceptual names used in the specification and is not represented by a
concrete type.

There were references to the renamed type Ruby (should have used
Runtime). These references have been fixed.