(PUP-3680) Specify Hash Type to not accept single V parameter
This updates the specification of the Hash Type access operation
that constructs a new Hash type instance. The convenience option
where a single type key was interpreted as a Hash with
a Scalar key, and given value type is now dropped as this proved
to be confusing in practice.

This commit also updates the "string to numeric conversion" section
to reflect that String should be converted to Numeric for arithmetic
operations, but not for comparisons. (This is how it is implemented
in 3.7.4/4.0.0). New decision is needed if this should change.