(PUP-3170) Cap Integer to 64 bit integer value
We do not handle BigInt values in puppet since serialization formats,
and database storage does not handle them without major pains.

This updates the specification to state that Integer is a 64 bit signed
integer value from -2^63 to 2^63-1.

Floating point is already "capped" in that it is a 64 bit "double
precision" value, and it simply degrades in precision. (No spec change
required).