(maint) Correct specification wrt. comparissons and numeric vs string
The specification had errors in that it described that comparissons ==,
!=, <, >, <=, >= to convert strings to numeric if both sides of the
comparisson could be converted to numeric.

That is not true. As a consequence of this, the text for the in operator
was updated.

Text is also added stating that if integer operations over or underflow
an error should be raised.