(maint) Clarify puppetserver log and pid locations
Without this patch the location of the puppetserver log files and pid
files are ambiguous.  The location inferred by the locations of the
agent.pid and puppet.log are problematic because puppetserver does not
run as root and as such will not be able to create new files in
`/var/run/puppetlabs` and `/var/log/puppetlabs`.

This pull request proposes puppetserver writable directories of
`/var/log/puppetlabs/puppetserver/` and
`/var/run/puppetlabs/puppetserver/`