(PUP-1806) Update specification of 4x func API for required/opt/repeated
The work on PUP-1806 concluded that required, optional and repeated
(varargs/capture) parameters should be declared with specific method
calls to:

* param (alias for required_param)
* required_param
* optional_param
* repeated_param
* block_param (alias for required_block_param)
* required_block_param
* optional_block_param

It was also decided to drop the `arg_count` method that was used to
control optionality, and capture/varargs.

This updates the specification (and fixes typos in the text).