(PUP-4194) Isolate puppet's logdir
When puppet executes as root, it sets permissions for file-based
settings to either `puppet:puppet` or `root:root` depending on whether
the puppet user and group exist or not.

Previously, puppet's `logdir` setting was `/var/log/puppetlabs`. If the
puppet user didn't exist, then puppet would set permissions to
root:root with mode 0750. If you later installed `puppetserver`, it
would create its log directory in `$logdir/puppetserver`, create the
puppet user and group, start the JVM as puppet, and fail to write to
its logfile, due to permissions on the grandparent `logdir`.

This commit changes the `logdir` to `/var/log/puppetlabs/puppet`, so
that puppet can manage permissions on its `logdir` without affecting
puppetserver, or other packages.

Note mcollective.log remains where it is.