(PUP-4520) Update case expression specification wrt. default entry
Before this the specification was not clear that the default entry (a
proposition having an option being a literal default) was skipped and
only selected last if no other option/proposition matched.

This commit updates the specification to clearly state that a literal
default option is skipped until the end, and it also makes clear that
literal defaults nsted inside of arrays and hashes are not considered to
be the case expression's default entry.