(maint) Remove unimplemented prosposals from puppet-functions.md
This removes several sections from the specification of functions
written in the puppet language that are not yet implemented and where
the text describes a proposal how they could be implemented.

The text is now updated to instead specifically point out what is not
supported yet: (getting and calling lambdas). Other features like
calling via a reference were simply deleted.

The feature that it should be allowed to reference default value
expressions via parameter references if the referenced value is on the
right is updated to reflect that this is currently undefined (reference
to ticket added).