(RE-6879) Update windows file paths
The Windows File paths have been changes as part of the Windows/Vanagon
release.

The rationale is to simplify user experience and maintenance on our
part. It makes our lives simpler because there's much less need for
branching in the component configs in puppet agent, and we don't have to
add 8 different locations to PATH or to RUBYLIB in order for our
software to find runtime dependencies

This work was originally handled under PR #63 as a maintenance update,
but is not being tracked under
[RE-6880](https://tickets.puppetlabs.com/browse/RE-6880).

This PR will remain open until Windows/Vanagon is released as part of
the Puppet-Agent release.