(PUP-5731) Specify that Puppet source is UTF-8 since Puppet 4.4.0
Before this, the specifiation said that only ASCII could be used and
that \uXXXX is the only way non ascii characters are guaranteed
to work across platforms (or by configuring Ruby to use UTF-8 as the
default). 

Now, it is clarly stated that UTF-8 is used, and that non ascii UTF-8
characters may appear in the source. 
The language for < Puppet 4.4.0 is updated as that still holds for older
versions (of the spec/puppet).