(PUP-7481) Specify Data and RichData
This specifies Data and RichData. It also cleans up the specification of
Hash and Array since the spec was behind in stating that unparameterized
Array and Hash types restrict entries to Data when they are in fact
unrestricted (Any).

The types `Object` and `TypeSet` are also added to the specification as
placeholders - mostly to make the RichData specification complete. The
two experimental types (`Object`, and `TypeSet` are not yet specified).