Change the Implementation to a plain class
By providing the implementation as a plain class, the complexity of the
runtime environment is gretly reduced. This makes testing much easier,
and removes a big chunk of magic from the stack.

The required follow-on change is to remove the `commands`, and `logger`
DSL from the provider environment. `commands` are replaced by a much
nicer explicit API based on the `childprocess` gem, enabling proper
background processing for those who need it, while preserving simple
calling conventions for common use-cases. The `logger` is replaced by a
libral-like `context` object that is passed through to `get`/`set`.