(PUP-9603) Update outdated specification of heredoc syntax tag
This updates the specification of the heredoc syntax tag:
`@(MARKER:<syntax>)`.

The text described the now deprecated experimental "binder" and how it
is used to extend the set of available syntax checkers. The binder was
dropped in Puppet 5 in favor of a much simpler mechanism. The
replacement mechanism (based on Puppet context) does not have the
ability to handle extensions (other than by modifying the source to push
additional things onto the Puppet context and this is not available to
users via configuration).

This commit updates the specification text to what is currently
supported and it enumerates the checkers available in puppet core.