(maint) Add smoke tests to validate JSON schema documents
Previously there was no validation on JSON schema documents in this repository.
Now that the Forge API directly links to this file, it is important that is at
the very least a valid JSON Schema document.

This commit adds a smoke test to validate that the tasks JSON schema documents
are syntatically valid.  This commit also adds a Travis CI configuration file
so that the tests can be run automatically.