Update RSAPI spec to match actual implementation
The implementation of the Resource API for the `simple_get_filter`
feature uses an empty array for the names paramter to signal that
all resources should be returned, not `nil`. Instead of changing
the code to match the spec, it probably makes more sense to
change the spec to match the implementation.