Add Module layout specification
This document describes all files/directories that are allowed or
mandatory within a module. The forge can use it as an allowlist and
filter other files out. r10k/code-manager can use it as an allowlist as
well when modules are pulled from version controll systems like github
(similar to the `--exclude-spec` option in r10k).

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>